ml-server/requirements.txt
confluent_kafka
google-cloud-language
google-cloud-translate==2.0.1
python-dotenv